requirements.txt
ymsteams
streamlit
AudioSegment
youtube_dl
ffmpeg
azure.cognitiveservices.speech
unicode_literals